% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findPara.R
\name{findPara}
\alias{findPara}
\title{Derive the parameters}
\usage{
findPara(train, bed, bim, fam, N1, N0)
}
\arguments{
\item{train}{train set after processed by 'extractInfo'}

\item{bed}{bed file after processed by 'extractInfo'}

\item{bim}{bim file after processed by 'extractInfo'}

\item{fam}{fam file after processed by 'extractInfo'}

\item{N1}{case number}

\item{N0}{control number}
}
\value{
A list including
estimated mu (muHat)
estimated sigma2 (sigmaHat2)
estimated proportion of non-associated SNPs (pi0)
estimated variance of effect sizes of associated SNPs (sigma02)
}
\description{
All the input files can be generated from 'extractInfo'
}
\references{
Song, S., Jiang, W., Hou, L. and Zhao, H. Leveraging effect size distributions to improve polygenic risk scores derived from genome-wide association studies. \emph{Submitted}.
}
\seealso{
\code{\link{extractInfo}}
}
\author{
Shuang Song, Wei Jiang, Lin Hou and Hongyu Zhao
}
