% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exdata.R
\docType{data}
\name{traindat}
\alias{traindat}
\title{Example data for training set}
\format{data.frame}
\usage{
data("traindat")
}
\description{
Summary statistics simulated in the manuscript Leveraging effect size distributions to improve polygenic risk scores derived from genome-wide association studies. Data from a QTL experiment on gravitropism in
}
\examples{
data("traindat")
 \dontrun{
 result=EBPRS(train=traindat, N1=364, N0=2063)}
}
\references{
Song S, Jiang W, Hou L, Zhao H (2020) Leveraging effect size distributions to improve polygenic risk scores derived from summary statistics of genome-wide association studies. PLoS Comput Biol 16(2): e1007565. https://doi.org/10.1371/journal.pcbi.1007565
}
\keyword{datasets}
