% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate}
\alias{validate}
\title{Validate the performance of EBPRS}
\usage{
validate(fam, score)
}
\arguments{
\item{fam}{fam file after processed by 'extractInfo'}

\item{score}{polygenic score generated by 'generateScore'}
}
\description{
Provide the performance evaluated by predictive r2 and AUC.
}
\examples{
validate(fam=matrix(sample(0:1,20*6,replace=TRUE),ncol=6),score=rnorm(20,0,1))
}
\references{
Song, S., Jiang, W., Hou, L. and Zhao, H. Leveraging effect size distributions to improve polygenic risk scores derived from genome-wide association studies. \emph{Submitted}.
}
\seealso{
\code{\link{extractInfo}}

\code{\link{findPara}}

\code{\link{generateScore}}
}
\author{
Shuang Song, Wei Jiang, Lin Hou and Hongyu Zhao
}
