% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate}
\alias{validate}
\title{Validate the performance of EBPRS}
\usage{
validate(score, truey)
}
\arguments{
\item{score}{polygenic score generated by 'EBPRS'}

\item{truey}{the true phenotype (the 6th column in the fam file)}
}
\description{
Provide the performance evaluated by predictive r2 and AUC.
}
\examples{
validate(score=rnorm(20,0,1), truey=sample(0:1,20,replace=TRUE))
}
\references{
Song S, Jiang W, Hou L, Zhao H (2020) Leveraging effect size distributions to improve polygenic risk scores derived from summary statistics of genome-wide association studies. PLoS Comput Biol 16(2): e1007565. https://doi.org/10.1371/journal.pcbi.1007565
}
\seealso{
\code{\link{EBPRS}}
}
\author{
Shuang Song, Wei Jiang, Lin Hou and Hongyu Zhao
}
